renamed `snake/` to `minesweeper/` for clarity
