+++ src/engine/redux/minesweeper/game/actions.ts
