+++ src/engine/redux/minesweeper/game/selectors.test.ts
