+++ src/engine/redux/minesweeper/models/Board.ts
