+++ src/engine/redux/minesweeper/models/Cell.ts
