+++ src/engine/redux/minesweeper/models/state.ts
