+++ src/engine/redux/minesweeper/reducers/gameReducer.ts
