+++ src/engine/redux/minesweeper/reducers/rootReducer.ts
