+++ src/engine/redux/minesweeper/services/SequenceGenerator.ts
