+++ src/engine/redux/minesweeper/store/mineSweeper.acceptance.spec.ts
