+++ src/engine/redux/minesweeper/store/root-action.ts
