+++ src/engine/redux/minesweeper/store/root-reducer.ts
